Update Full Data Log Reg
Fixed error with the logistic regression code that was using all of the data. Also made adjustments to the hyperparameter tuning which now uses f1 precision for scoring rather than accuracy. This along with using more students for validation helped make a more manageable cost history for XGBoost Classifier that can be fixed with the right parameters.